Fix bugs found during live MDbg cross-validation testing
- Debugger_1.cs: Fix pointer dereference for g_pDebugger global (was
  missing ReadPointer to dereference the pointer-to-pointer)
- DacDbiImpl.FlushCache: Call _target.Flush() to clear ProcessedData
  cache, return S_OK instead of E_NOTIMPL when no legacy
- DacDbiImpl.GetAssemblyFromDomainAssembly: Fix to dereference
  DomainAssembly at offset 0 to get Assembly* (was incorrectly
  treating DomainAssembly as Module*)
- DacDbiImpl.GetModuleForDomainAssembly: Fix to chain through
  DomainAssembly -> Assembly -> Module via GetModuleHandleFromAssemblyPtr
- DacDbiImpl.HasTypeParams: Fall back to legacy - native uses
  ContainsGenericVariables() which has no cDAC contract equivalent
  (was incorrectly using HasTypeParam() which checks different semantics)
- dacdbiimpl.cpp: Add native DBI shimming to create cDAC DacDbi
  interface when ENABLE_CDAC=1
- cdac.h/cpp: Add CreateDacDbiInterface entry point
- cdac_reader.h: Add cdac_reader_create_dacdbi_interface declaration
- Test fixes: uint for GetGCHeapCount, skip GetCurrentException
  gracefully when GetThrowableObject is not implemented

Verified with MDbg Checked cross-validation (ENABLE_CDAC=1):
  Sanity: 2/2 pass, Stepping: 15/15 pass, Breakpoints: 3/3 pass

+++ src/native/managed/cdac/tests/DumpTests/DacDbi/DacDbiGCDumpTests.cs
@@ -79,7 +79,7 @@ public class DacDbiGCDumpTests : DumpTestBase
         bool contractValid = gc.GetGCStructuresValid();
         Assert.Equal(contractValid, heapInfo.areGCStructuresValid == Interop.BOOL.TRUE);
 
-        int heapCount = gc.GetGCHeapCount();
-        Assert.Equal((uint)heapCount, heapInfo.numHeaps);
+        uint heapCount = gc.GetGCHeapCount();
+        Assert.Equal(heapCount, heapInfo.numHeaps);
     }
 }
